+++ src/synthesizer/extensions/doppler_particle_spectra.cpp
@@ -214,7 +214,7 @@ static void shifted_spectra_loop_cic_omp(GridProps *grid_props,
 
     /* Split the work evenly across threads (no single particle is more
      * expensive than another). */
-    int nparts_per_thread = parts->npart / nthreads;
+    size_t nparts_per_thread = parts->npart / nthreads;
 
     /* What thread is this? */
     int tid = omp_get_thread_num();
@@ -467,7 +467,7 @@ static void shifted_spectra_loop_ngp_omp(GridProps *grid_props,
 
     /* Split the work evenly across threads (no single particle is more
      * expensive than another). */
-    int nparts_per_thread = parts->npart / nthreads;
+    size_t nparts_per_thread = parts->npart / nthreads;
 
     /* What thread is this? */
     int tid = omp_get_thread_num();
